# services/aetherus-orbital/openapi/aetherus-v2.openapi.yaml
openapi: 3.1.0
info:
  title: Aetherus V2 API
  description: SPACE + CONTROL + ORBIT + INTELLIGENCE local product runtime. Scientific validation states are explicit.
  version: 0.5.0
paths:
...
  /internal/ingestion/runs:
    get:
      summary: Ingestion Runs
      operationId: ingestion_runs_internal_ingestion_runs_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/objects/{object_id}:
    get:
      summary: Object Detail
      operationId: object_detail_v1_objects__object_id__get
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Object Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/evidence/{evidence_id}:
    get:
      summary: Evidence Detail
      operationId: evidence_detail_v1_evidence__evidence_id__get
      parameters:
      - name: evidence_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Evidence Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/states/{state_id}:
    get:
      summary: State Detail
      operationId: state_detail_v1_states__state_id__get
      parameters:
      - name: state_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: State Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/snapshots/{snapshot_id}:
    get:
      summary: Snapshot Detail
      operationId: snapshot_detail_v1_snapshots__snapshot_id__get
      parameters:
      - name: snapshot_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Snapshot Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/graph/{node_id}:
    get:
      summary: Graph Detail
      operationId: graph_detail_v1_graph__node_id__get
      parameters:
      - name: node_id
        in: path
        required: true
        schema:
          type: string
          title: Node Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/intelligence/events/{event_id}:
    get:
      summary: Intelligence Event
      operationId: intelligence_event_v1_intelligence_events__event_id__get
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/universe:
    get:
      summary: Universe Get
      operationId: universe_get_v1_universe_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    patch:
      summary: Universe Patch
      operationId: universe_patch_v1_universe_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UniversePatch'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/space/state:
    get:
      summary: Space State
      operationId: space_state_v1_space_state_get
      parameters:
      - name: at
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: At
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/space/ephemeris:
    get:
      summary: Ephemeris
      operationId: ephemeris_v1_space_ephemeris_get
      parameters:
      - name: target
        in: query
        required: false
        schema:
          type: string
          default: EARTH
          title: Target
      - name: at
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: At
      - name: observer
        in: query
        required: false
        schema:
          type: string
          default: SUN
          title: Observer
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/missions/{mission_id}:
    get:
      summary: Mission Detail
      operationId: mission_detail_v1_missions__mission_id__get
      parameters:
      - name: mission_id
        in: path
        required: true
        schema:
          type: string
          title: Mission Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/missions/{mission_id}/replay:
    get:
      summary: Mission Replay
      operationId: mission_replay_v1_missions__mission_id__replay_get
      parameters:
      - name: mission_id
        in: path
        required: true
        schema:
          type: string
          title: Mission Id
      - name: at
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: At
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/missions/{mission_id}/handover:
    get:
      summary: Mission Handover
      operationId: mission_handover_v1_missions__mission_id__handover_get
      parameters:
      - name: mission_id
        in: path
        required: true
        schema:
          type: string
          title: Mission Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/orbit:
    get:
      summary: Orbit State
      operationId: orbit_state_v1_orbit_get
      parameters:
      - name: at
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: At
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/conjunctions:
    get:
      summary: Conjunctions
      operationId: conjunctions_v1_conjunctions_get
      parameters:
      - name: at
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: At
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/archive/search:
    get:
      summary: Archive Search
      operationId: archive_search_v1_archive_search_get
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/archive/time-machine:
    get:
      summary: Archive Time Machine
      operationId: archive_time_machine_v1_archive_time_machine_get
      parameters:
      - name: at
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: At
      - name: mode
        in: query
        required: false
        schema:
          type: string
          default: ARCHIVED_STATE
          title: Mode
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/scenarios/run:
    post:
      summary: Scenario Run
      operationId: scenario_run_v1_scenarios_run_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/llm/explain:
    get:
      summary: Llm Explain
      operationId: llm_explain_v1_llm_explain_get
      parameters:
      - name: locale
        in: query
        required: false
        schema:
          type: string
          default: en
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/briefings/current:
    get:
      summary: Briefing Current
      operationId: briefing_current_v1_briefings_current_get
      parameters:
      - name: locale
        in: query
        required: false
        schema:
          type: string
          default: en
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/search:
    get:
      summary: Search
      operationId: search_v1_search_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          title: Q
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/scene/{mode}:
    get:
      summary: Scene
      operationId: scene_v1_scene__mode__get
      parameters:
      - name: mode
        in: path
        required: true
        schema:
          type: string
          title: Mode
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/subscription/capabilities:
    get:
      summary: Capabilities
      operationId: capabilities_v1_subscription_capabilities_get
      parameters:
      - name: plan
        in: query
        required: false
        schema:
          type: string
          default: FREE
          title: Plan
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/protect/candidates:
    post:
      summary: Protect
      operationId: protect_v1_protect_candidates_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProtectRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProtectRequest:
      properties:
        protected_object_id:
          type: string
          title: Protected Object Id
        protected_active:
          type: boolean
          title: Protected Active
          default: true
        candidates:
          items:
            type: object
          type: array
          title: Candidates
      type: object
      required:
      - protected_object_id
      - candidates
      title: ProtectRequest
...
    ScenarioRequest:
      properties:
        kind:
          type: string
          title: Kind
        target_object_ids:
          items:
            type: string
          type: array
          title: Target Object Ids
        protected_object_ids:
          items:
            type: string
          type: array
          title: Protected Object Ids
        parameters:
          type: object
          title: Parameters
        assumptions:
          items:
            type: string
          type: array
          title: Assumptions
        seed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Seed
      type: object
      required:
      - kind
      - target_object_ids
      title: ScenarioRequest
...
    UniversePatch:
      properties:
        current_time_utc:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Current Time Utc
        selected_object:
          anyOf:
          - type: string
          - type: 'null'
          title: Selected Object
        selected_event:
          anyOf:
          - type: string
          - type: 'null'
          title: Selected Event
        camera_focus:
          anyOf:
          - type: string
          - type: 'null'
          title: Camera Focus
        space_scale:
          anyOf:
          - type: string
          - type: 'null'
          title: Space Scale
        active_mission:
          anyOf:
          - type: string
          - type: 'null'
          title: Active Mission
        active_orbital_shell:
          anyOf:
          - type: string
          - type: 'null'
          title: Active Orbital Shell
        scenario_state:
          anyOf:
          - type: object
          - type: 'null'
          title: Scenario State
        time_mode:
          anyOf:
          - type: string
          - type: 'null'
          title: Time Mode
      type: object
      title: UniversePatch
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...